Added support for conditionally skipping tests using CPP.
Although it is easy enough to surround a unit test with an #ifdef to
prevent compilation, the automatically generated suite will still
include a call to the now non-existent procedure.

The driving use case is for tests that should be skipped when not
compiling with MPI, so for now only @mpitest have the new syntax,
which is of the form:

@mpitest(npes=[...], ifdef=<TOKEN>)

In the suite factory there will now be a
#ifdef <TOKEN>
...
#endif

that prevents appending the test to the suite unless <TOKEN> is true.


If anyone has a suggestion for a better syntax/name for this let me
know.  Also if you need it to work for non-mpi tests, just say so.