Added new preprocessor macros "@mpiAssert*" and associated framework support logic.
@mpiAssert* behaves much the same as @assert*, but will return to the
calling procedure if _any_ process has raised an exception.  This is
important in situations where only some (but not all) processes
register an exception.  If the original @assert is used, only a subset
of processes will continue through the remaining logic within a test,
and will then hang if there is a subsequent MPI synchronization point.

Although unit tests are in place, I would consider this feature to be
"beta" until it has been vetted in some real examples.

Many thanks to Matthew Hambley for pushing for a proper solution to
this issue.