Preliminary attempt to support CMake for building pfunit.
Note that I expect to always maintain a conventional GNU makefile for
this project to minimize dependencies for new users.  However, I've
grown to like CMake to the point that I want to use it for my own
routine development.

This works in my environment with ifort and gfortran. 

To Do:
   - allow to work with MPI (should be easy)
   - add support for Windows (no relevant experience in house)
   - add support for NAG (should be easy)
   - add support for AIX (should be easy)