Add ROBUST option.
The ROBUST option is on by default for Unix and off for Windows. Turning it
off removes RobustRunner and UnixProcess. Some conditionals testing for
Windows are changed to use this option.

The driver also has a BUILD_ROBUST CPP Macro now; the user must define it
to enable the robust runner (in addition to supplying the run-time option).

The default behavior of the pFUnit build and the driver is now a bit
inconsistent. It may make sense to have the ROBUST mode off by default in
both places in the future, or else to make a modification to remove the
need for the CPP macro from the driver.