Make OpenMP run failure fatal again.
An earlier change to FindOpenMP_Fortran.cmake made it a WARNING instead of
a FATAL_ERROR for an OpenMP executable to fail to run.

This change misunderstood the nature of an issue on Mira, which was really
due to the fact that a bare "-D" is not correct for defining a CPP Macro
for the XLF compiler. This was fixed by moving the definition into
COMPILE_DEFINITIONS in a different way, allowing CMake to fix the flag
automatically.

Once this was done, a run failure could be made a fatal error again. (Note
that if this really was a cross-compilation issue, this should have been
signalled using the CMAKE_CROSSCOMPILING variable anyway.)