Eliminated "Semi-Automatic" subdirectories from Examples directory.
In pFUnit 1.0 the "wrapping" mechanism was fragile, which motivated
supporting a manual and a preprocessor mechanism for implementing user
tests.  The Examples directory was created with that dual mechanism in
mind.  However, since we've never created the manual examples and the
preprocessor seems reasonably successful/robust, the extra complexity
is just silly.

If we ever create a manual example (i.e. explicit calls to the
framework instead of @test, @assert, etc.) it would not need to be
duplicated for every example.