pFUnit now attempts to return failure when considered as a Unix application.
Previous versions of pFUnit returned 'success' (i.e. 0) when run as a
Unix application even when some tests failed.  This change will make
it easier to embed unit testing within other scripts - e.g. for
regression testing.   

This fix should work for serial builds on any system that supports the
Fortran 2008 "error stop" statement.  (And we should be able to use
#ifdef for vendor specific mechanisms on other compilers - e.g. NAG 5.3.)

Unfortunately, it is difficult to guarantee behavior for arbitrary MPI
flavors.  The modification will probably work on most if not all
variants, but it is not guaranteed.  Users are encouraged to verify
before relying on this feature.

NOTE: This modification exposed a bug in GCC 4.8.2.  A patch has been
found and will appear in 4.8.3 and 4.9.0.  If you need to use older
versions of gfortran you should use the 'pfunit_2.1.x' branches from
the repository.