Fix for parameterized extensions of MpiTestCase.
Old usage of setNumProcesses() was not causing any problems with
existing tests, but was breaking realistic examples because the base
class TestParameter component was being doubly allocated.  Once in the
user-defined constructor, and then again with setNumProcesses().

Fix was to eliminate the method for MpiTestCase, and only use it on
MpiTestParameter.