Separate invocation of CC for preprocessing is no longer necessary.
Previously the stringify operation was not supported by FPP under NAG,
but that is apparently working now.  May need to restore the
capability for backward compatibility, but want to push this to
development for further testing in that direction.

Various cleanup in selfTests.F90 was done while examining this issue.