More fixes for misuse of PRESENT.
Thanks to a comment from Sean Santos, pFUnit now builds with
"-C=present" for the NAG compiler which should enable us to prevent
re-introducing similar bugs in the future.

This flag identified a few more instances to fix.

This commit also undoes some changes in the previous commit that were
unnecessary as the OPTIONAL arguments "location" and "message" were
being passed along to other procedures for which those arguments are
optional.  (New reason to dislike optional arguments - impossible for
a casual reader to know if a PRESENT() check is required.)