Modified name of "wrapper" module when using @suite.
Motivation was to help avoid very long module names for a project
which is autogenerating code with long names.  @suite allowed a custom
name for the user-visible wrapper function, but had no impact on the
user-invisible wrapper module.

This change should not impact any users, but if I'm wrong, I will
revert and add a separate option to specify the wrapper module name.

New convention is Wrap<suiteName>.   Old convention is Wrap_<baseFileName>_mod.