Added alternative option for specifying access to test suites defined in a module function.
When developing generic tests (e.g. overloaded for different types),
it can be inconvenient to use the provided python preprocessor.  In
these cases, it is often easier/better to place the manually
constructe test suite generator inside of a module.  In that case
ADD_TEST_SUITE(s) is inadequate.

The new macro ADD_MODULE_TEST_SUITE(m,s) allows the specification of a
suite generator named <s> that is USE'd from module <m>.