Implementing better protection of namespace.
All pFUnit modules now have "PF_" as a prefix for their names.

This eliminates the need for pFUnitPackage which is now deleted.
Likewise the module "pFUnit_mod" is now simply "pFUnit" to indicate
that it is a package rather than a mere module.

Note: names for classes and files have not changed.