Implemented simple glob pattern matching facility.
Cheated - found implementation in C online and did a loose
translation.  I sort of understand how it is working, but would like
to understand it better and eliminate recursion.  Unfortunately, the
examples without recrusion use GOTO.

Note that speed is not expected to ever be an issue for the glob cases
used in pfunit.