Just a bit of cleanup.
There are some issues in this file that hardwire which directory the
tests must run in.  This was noticed while introducing CTest which
tries to run in a different directory than that used for the legacy
'make tests'.

Did not fix this issue, but did manage to add a deletion for the tmp
file created in one of the tests.