Added support for @ignore test annotation.
This unfortunately affected about a dozen classes -- even though the
ignoreTest() method is a currently a noop in the various TestRunner
classes.

Things will get a bit more complicated when more general @rule
annotations are included.

This also pressed the issue of needing TestSurrogate which is adding
complexity for little value.  The original rationale was to avoid
putting multiple base classes into a single Fortran module.  But ...
time to compromise a bit on that I think.

Added proper test for @ignore functionality.