Fix for formatted Infinity.
NAG 6.2 uses "Inf" for formatted infinity rather than "Infinity" with
6.1.  This was causing some tests to fail due to incorrect
expectations.  The fix was to write a helper function that returns a
string that is formatted infinity for whatever compiler is involved.
Tests can then set expected assertion failure messages using that
string rather than a litereal 'Infinity'.

I suspect F2008 introduced a requirement for the format of Inf and
this is why the behavior of NAG changed.