Workaround for Gfortran bug.
GFortran EXECUTE_COMMAND_LINE does not behave correctly when EXITSTAT
is used to detect whether a system has xmllint.  No one has complained
about this in Test_XmlPrinter in 4 years time so apparently xmllint is
present, but by default it is not there in Travis.

More robust workaround would be to test for xmllint during cmake
config.