Fixed bug in XML unit test.
This bug has not caused problems because apparently all relevant
systems have xmllint installed.  The intent was to silently skip the
xmllint check when unavailable, but the wrong status flag was used on
Fortran's EXECUTE_COMMAND_LINE.

The problem was exposed when testing under travis which has a more
distilled environment.  A patch earlier today added xmllint to travis.
In theory it is no longer needed but it is probably better to allow
the extra check during automated testing.