Implemented pervasive templating to follow Java's design.
The design is becoming somewhat debatable when considering the
limitations of Fortran. (No true templates.)     I think the next step
is to change the templates from being self-contained in terms of their
supported types and into something which a file is generated
per-type.   This would allow users to template their own types.