Lots more progress on hamcrest features.
Have introduced meta-matchers:
   - TypeSafeMatcher
   - FeatureMatcher
   - DescribedAs

TypeSafeMatcher is not quite as useful as it is in Java where the
subclass can assume arguments are of the necessary type (and thereby
skip SELECT TYPE).  But it at least allows the main methods, matches()
and describe_mismatch(), to only have the one clause for SELECT TYPE
without a CLASS DEFAULT clause.