Introduced Surrogate Pattern.
The interface for append_list() really needs to be a gFTL Vector of
polymorphic type.  This prevented the previous kludge of placing
SelfDescribing and BaseDescription in the same module.

Now, the Surrogate Pattern is used to split the two classes into
separate modules.

Work should now begin on modifying the append_list() interface and
implementation.