Fixed a few important deficiencies in new assert templates.
Apparently I had forgotten to go back and create an include file that
USEs all of the generated asserts for various ranks in the top Asssert
module.  The tests generally directly used the specific rank modules,
which is why this was not detected sooner.

This commit also now includes some tests for ComplexAssert.  These are
hardly complete, but the code may be ok, as it is derived from the
Real case.