CMake support for projects that use CMake.
Created a function add_pfunit_ctest() which takes a number of keyword
arguments and attempts to build an executable and ctest test from a
collection of test sources and additional sources.

Also fixed a few minor issues with the install step that were exposed
when attempting to exercise this new capability.

Note: this version should be considered preliminary.  More options are
needed, and not all existing options have been tested (e.g. MPI).