Fixes #124 - improved support for gmake.
The ticket indicated that base.mk is missing in pFUnit 4.0.
Investigation shows that this is true, but that it is replaced
now by PFUNIT.mk.

But while there, I noticed that support for MPI had not been properly
included.  This is complete now, and the MPI demo (other repo) has
been updated accordingly.