Fixes #154 - EXTERNAL required
The new approach in pFUnit 4.0 for the driver, treats optional
initialize() and finalize() as procedure pointer targets via
PFUNIT_EXTRA_INITIALIZE and PFUNIT_EXTRA_FINALIZE.  Within driver, a
procedure pointer is declared with an implicit interface, but this
approach requires that the target procedure have the EXTERNAL
attribute or be accessed through a module.

Unfortunately, one cannot simply always add EXTERNAL, because that it
conflicts if the user also uses PFUNIT_EXTRA_USE.   So users must not
mix and match.   Either the INITIALIZE and FINALIZE methods are given
through a module or they are external.

empty