Fixes #180.
The constructor for TestMethod now allows setup without teardown and
vice-versa.  The othe overload for this constructor with non-optional
dummy arguments is deprecated, and the dummy argument spelling has
changed to avoid conflicts.  Users should obtain identical results
regardless of whether they use keyword association or not.   But it
is a bit subtle as to why.