Quick and dirty AssertEqual for string arrays
The generated errors are not particulary informative, as this
 implementation just iterates over individual strings.

The whole StringAssert package needs serious refactoring.  I fixed a
bit with this PR, but really need to isolate a procedure for comparing
the string and then drive it from different ranks.  Then we can get an
informative message out about _which_ string is the first one that is
different and when which character is the first that is different.